feat: add styles for new elements on the table
